+++ src/app/courses/page.tsx
@@ -1,26 +1,12 @@
-import rankingDataRaw from "../../../data/data_files/processed/ranking_courses.json";
-import { Ranking, RankingItemRaw } from "../components/Ranking";
+import { Suspense } from "react";
+import CourseRanking from "./courseRanking";
 
-const rankingData = JSON.parse(JSON.stringify(rankingDataRaw));
-const rankingItems: RankingItemRaw[] = [];
-
-for (const k in rankingData) {
-    const v = rankingData[k];
-    rankingItems.push({
-        title: k,
-        score: v.course_mean,
-        nr: v.num_response,
-        link: "/courses/" + k.slice(0, 4) + k.slice(5),
-    });
-}
-
-export default function CourseRankings() {
+export default function Page() {
     return (
-        <div>
-            <h1 className="hidden">Course Rankings on SFQ Scores</h1>
-            <Ranking items={rankingItems} searchPrompt="Search for courses here..." scoreName="course_mean" />
-        </div>
-    );
+        <Suspense fallback={<div>Loading...</div>}>
+            <CourseRanking />
+        </Suspense>
+    )
 }
 
 export async function generateMetadata() {
