+++ src/app/instructors/page.tsx
@@ -1,27 +1,12 @@
-import rankingDataRaw from "../../../data/data_files/processed/ranking_instructors.json";
-import { Ranking, RankingItemRaw } from "../components/Ranking";
-import { getNameByItsc } from "../utils";
+import { Suspense } from "react";
+import InstructorRanking from "./instructorRanking";
 
-const rankingData = JSON.parse(JSON.stringify(rankingDataRaw));
-const rankingItems: RankingItemRaw[] = [];
-
-for (const k in rankingData) {
-    const v = rankingData[k];
-    rankingItems.push({
-        title: getNameByItsc(k),
-        score: v.instructor_mean,
-        nr: v.num_response,
-        link: "/instructors/" + k,
-    });
-}
-
-export default function InstructorRankings() {
+export default function Page() {
     return (
-        <div>
-            <h1 className="hidden">Instructor Rankings on SFQ Scores</h1>
-            <Ranking items={rankingItems} searchPrompt="Search for instructor here..." scoreName="instructor_mean" />
-        </div>
-    );
+        <Suspense fallback={<div>Loading...</div>}>
+            <InstructorRanking />
+        </Suspense>
+    )
 }
 
 export async function generateMetadata() {
